backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pgvector==0.2.5
sentence-transformers==3.0.1
dramatiq==1.16.0
redis==5.0.4
pyyaml==6.0.1
pymilvus==2.4.0
marshmallow==3.19.0
openai>=1.50.0